Move formatting to `format/format.rs`
